made changes based on comment
