+++ frontend/src/pages/AddTaskPage.module.css
@@ -1,0 +1,20 @@
+.starting {
+    margin: 3rem auto;
+    text-align: center;
+  }
+  
+  .starting h1 {
+    font-size: 2rem;
+  }
+  
+  .starting img {
+    align-items: center;
+    margin: 1rem auto;
+    width: 80%;
+    height: auto;
+  }
+  
+  .warning {
+    color: #FF6347;
+  }
+
